Make code dry
